first model trained :)

+++ training/MathExpressionDataset.py
@@ -55,7 +55,7 @@ class MEdataset(object):
             
         # convert everything into a torch.Tensor
         boxes = torch.as_tensor(boxes, dtype=torch.float32)
-        labels = torch.as_tensor(labels, dtype=torch.float32)
+        labels = torch.as_tensor(labels, dtype=torch.int64)
         image_id = torch.tensor([idx])
         
         area = (boxes[:, 3] - boxes[:, 1]) * (boxes[:, 2] - boxes[:, 0])
